% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_proper.R
\name{lm_proper}
\alias{lm_proper}
\title{Fit a Bayesian linear model with Normal-Gamma prior via Gibbs Sampling.}
\usage{
lm_proper(data, formula, burnin = 5000, iterations = 5000, thin = 1,
  beta_prior_mean, beta_prior_cov, tau_prior_shape, tau_prior_rate,
  start_beta)
}
\arguments{
\item{data}{A data frame containing the variables to be used in the model.}

\item{formula}{A formula object describing the model fit.}

\item{burnin}{Number of burn in iterations}

\item{iterations}{Number of MCMC iterations}

\item{thin}{Number of thinning iterations}

\item{beta_prior_mean}{Mean vector for the prior on \eqn{\beta}.}

\item{beta_prior_cov}{Covariance matrix for the prior on \eqn{\beta}.}

\item{tau_prior_shape}{Shape parameter for the prior on \eqn{\tau}.}

\item{tau_prior_rate}{Rate parameter for the prior on \eqn{\tau}.}

\item{start_beta}{Beta starting vector. The default is \eqn{\beta}-hat.}
}
\value{
A data frame containing MCMC samples from the posterior distribution.
}
\description{
Fits a Bayesian linear regression model with a Normal prior on beta and a
Gamma prior on the precision tau. Model fitting is done via a Gibbs sampler.
}
