% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm-improper.R
\name{lmm_improper}
\alias{lmm_improper}
\title{Fit a Bayesian Generalized Linear Mixed Model With Improper Prior}
\usage{
lmm_improper(data, formula, burnin = 5000, iterations = 5000,
  thin = 1, lambda_prior_shape, lambda_prior_rate, start_theta)
}
\arguments{
\item{data}{A data frame containing the variables to be used in the model.}

\item{formula}{A formula object describing the model fit.}

\item{burnin}{Number of burn in iterations}

\item{iterations}{Number of MCMC iterations}

\item{thin}{Number of thinning iterations}

\item{lambda_prior_shape}{A length 2 vector containing the shape for the prior
on \eqn{\lambda_e} and \eqn{\lambda_u}, in that order.}

\item{lambda_prior_rate}{A length 2 vector containing the rate for the prior
on \eqn{\lambda_e} and \eqn{\lambda_u}, in that order.}

\item{start_theta}{Starting vector for \eqn{\theta = (\beta' u')'}.
Default is the frequentist guess.}
}
\value{
A list containing MCMC samples from the posterior distribution of
 \eqn{\beta}, \eqn{u}, \eqn{\sigma_e}, and \eqn{\sigma_u}.
}
\description{
Fists a Generalized Linear Model with a flat prior on \eqn{\beta} and
improper Gamma priors on the precision for \eqn{u} and \eqn{e}.
}
