% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm-proper.R
\name{lmm_proper}
\alias{lmm_proper}
\title{Fit a linear mixed model with a proper prior.}
\usage{
lmm_proper(data, formula, burnin = 5000, iterations = 5000, thin = 1,
  beta_prior_mean, beta_prior_cov, lambda_prior_shape, lambda_prior_rate,
  start_theta)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{A formula describing the model fit. Passed to
\code{\link[lme4:lmer]{lme4::lmer}} to construct model matrices.}

\item{burnin}{Number of burn in iterations}

\item{iterations}{Number of sampling iterations}

\item{thin}{Number of thinning iterations}

\item{beta_prior_mean}{(Optional) Mean vector for the prior on beta.
Defaults to zero.}

\item{beta_prior_cov}{(Optional) Covariance matrix for the prior on beta.
Defaults to \code{diag(p) * 100} where p is the number of predictors.}

\item{lambda_prior_shape}{(Optional) Shape parameter for the prior on the
precision of the error term and random effects, in that order. Defaults
to \code{c(0.001, 0.001)}.}

\item{lambda_prior_rate}{(Optional) Rate parameter for the prior on the
precision of the error term and random effects, in that order. Defaults
to \code{c(0.001, 0.001)}.}

\item{start_theta}{(Optional) Starting vector for
\eqn{\theta = (\beta' u')'}, the concatenation of the fixed and random
effects coefficients. Defaults to the frequentist estimate.}
}
\value{
A list containing MCMC samples from the posterior distributions of
 the fixed effects, random effects, and standard deviations for the error
 term and fixed effects.
}
\description{
Fits a Bayesian linear mixed model with a normal prior on the fixed effects
    beta and an gamma prior on the precision lambda for the fixed effects
    and error term. The formula syntax is the same as
    \code{\link[lme4:lmer]{lme4::lmer}} which is used to obtain the model
    matrices. Only a single intercept is supported for the random effects.
}
\section{Warning}{

 We leave it to the user to ensure only a single intercept is specified
    in the model formula.
}

\references{
Roman, J. C. and Hobert, J. P. (2015). Geometric ergodicity
    of Gibbs samplers for Bayesian general linear mixed models with
    proper priors. Linear Algebra and its Applications,47 354–77.
}
