% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_proper.R
\name{lm_proper}
\alias{lm_proper}
\title{Fit a linear model with a proper prior.}
\usage{
lm_proper(data, formula, burnin = 5000, iterations = 5000, thin = 1,
  beta_prior_mean, beta_prior_cov, tau_prior_shape, tau_prior_rate,
  start_beta)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{A formula describing the model fit. Passed to
\code{stats::lm()} to construct the model matrix.}

\item{burnin}{Number of burn in iterations.}

\item{iterations}{Number of sampling iterations.}

\item{thin}{Number of thinning iterations.}

\item{beta_prior_mean}{(Optional) Mean vector for the prior on beta. Defaults
to zero.}

\item{beta_prior_cov}{(Optional) Covariance matrix for the prior on beta.
Defaults to \code{diag(p) * 100} where p is the number of predictors.}

\item{tau_prior_shape}{(Optional) Shape parameter for the prior on tau.
Defaults to 0.001.}

\item{tau_prior_rate}{(Optional) Rate parameter for the prior on tau.
Defaults to 0.001.}

\item{start_beta}{(Optional) Beta starting vector. Defaults to beta hat,
the frequentist estimate.}
}
\value{
An object of class \code{geblm} containing samples from the
    posterior distributions of beta and sigma.
}
\description{
Fits a Bayesian linear regression model with a normal prior on beta and a
    gamma prior on the precision tau. 'Wide' priors are used by default.
}
\examples{
fit <- lm_proper(mtcars, mpg ~ wt)

}
